projects/blackjack_v_0.1_project/requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
PyJWT==2.1.0
Werkzeug==2.0.1
WTForms==3.0.0
SQLAlchemy==1.4.46
Jinja2==3.0.1
itsdangerous==2.0.1
python-dotenv==0.19.0
email-validator==1.1.3
alembic==1.7.3
blinker==1.4
pytest==7.4.3